back/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
exceptiongroup==1.3.0
executing==2.2.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==25.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pywin32==310
pyzmq==26.4.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.21.0
